# Fix performance issues when having remote schema with multiple refs