@sirosen 2024-06-27
The `CacheDownloader` used inside of the ref resolver does not have `disable_cache=...` set on init. As a result, it uses the class default. i.e. `self._disable_cache` is always `False` in the new usage site, even if `--no-cache` was passed.

> But that does not quite work when one wants to cache multiple files. Like one might need to somehow use the complete URL to avoid those conflicts

The issue in which I outlined a solution discusses this. By taking the full URI and using it to build a filename, I think we can solve this problem.

> I also wondered if one might just want to use some in-memory cache. Like there should not be a need to re-read the file from disk.

I'm slightly confused by this comment, given that this PR is about adding on-disk caching.
The ref resolution mechanism provided by `referencing` is already holding results in memory.

> As a sidenote: the current caching does not prevent requests, it is only for writing the file on disk? Like some webservers set cache headers that might be leveraged.

This seems inaccurate to me? Or if we are doing downloads of the schema even on a cache hit, it would mean there is a bug.
The caching usage today reads the `Last-Modified` header and compares against the mtime on disk.
There are lots of other headers (e.g. E-Tags) which could be used, but this was the simple 95% effective solution I started with.

Are you seeing redownloads of schemas even when there is a `Last-Modified` header provided? A reproducer would be key to tracking down this sort of bug.