@sirosen 2024-06-27
It hit me as soon as I posted that last comment that there is a bug and I even know when/how it got introduced.
https://github.com/python-jsonschema/check-jsonschema/issues/453 captures it, but the caching got borked when I added a safety check that the remote content is parseable -- I added it incorrectly and didn't catch it during testing.

I can't work on a fix for it right now, during my workday, but I'll be circling back on it as soon as I have time.