@sirosen 2024-07-01
I would advise against working on this PR.
I have a complete implementation I worked up at this point, but want to make some minor changes for non-JSON refs (e.g. YAML files being used). I'll have it in a PR shortly and plan to close this and #452 once it's merged and released.