@sirosen 2024-07-08
I'm closing this as superseded by #454, which should deliver these gains. This work was the original inspiration and it showed a lot of what had to happen, but I was unfortunately not able to structure the changeset in a way that could meaningfully be based off of this commit. I did, however, credit you in the changelog.

This will be released in v0.29.0 in just a few minutes time. Thanks again for the contribution!