@alex1701c 2024-07-09
Thank you for your efforts and credeting me!

However, even with the latest version the performance is still very slow. Like if I pass in the same list of files, the schemas get requested lots of times. The latency still adds up to around 30 seconds. I think this is due to determining if we have a cache hit, we still make the request.