@sirosen 2024-07-09
To determine if there is a cache hit, we need to open a connection and get some bytes back, to get the headers. Are you saying that the behavior you see is that there's a full download? It's possible that there's some case that I missed during testing which causes this.

It might be best to open an issue with a couple of example files to demonstrate the issue. When I tested things out after the improvements, I saw the behavior I was expecting, which showed a measurable improvement.