@alex1701c 2024-07-10
But if we have revalidated/downloaded the same file in the same process, there is no need to open a HTTP connection again. My original issue was not about the data transfered, but the amount of requests that add up to quite a slow linting.