@sirosen 2024-07-10
There is an in-memory cache which is used [here](https://github.com/python-jsonschema/check-jsonschema/blob/8a0cfeeb5e8913eec54e413049181609083cd6ba/src/check_jsonschema/schema_loader/resolver.py#L82) to retrieve data when ref resolution requests the same URL twice. This is built once per validator, and only one validator _should_ be built per process.

I'm very much open to the idea that there's some bug which causes this not to function correctly, but I need more information to reproduce in order to track it down. Can you please include some sample files which demonstrate the issue? I have the schema from your initial PR text, but I don't know what the content is supposed to be, so it's difficult for me to understand what you're seeing.