@alex1701c 2024-07-13
But create_retrieve_callable gets called multiple times in case we have mutliple files passed in?

The example repo I have used is https://invent.kde.org/network/kdeconnect-kde/