@sirosen 2024-07-15
> But create_retrieve_callable gets called multiple times in case we have mutliple files passed in?

Are you certain of this? That should be called once per Validator instantiation, and therefore once per top-level schema, i.e. once per invocation. If it is being called for each `$ref`, that would be a bug which needs resolution.